Allow personal series to be transferred from a user's T-account to their staff account
Email addresses are used as lookup when searching for personal series.
It seem to be standard policy that whenever a contract staff member becomes a
permanent staff member, their email address stays the same. Thus, new series are
not created for staff members who initially used their T-accounts, but now use
their UCT accounts. This occurrence is understandably confusing to end users
who progress through UCT's ranks.

A new endpoint endpoint has been added, namely /mail/taccounts. POSTing to
this endpoint will
- fetch all personal series belonging to T-accounts,
- check whether their owner has an associated UCT account,
- and if so, forward an email with links to transfer their personal series if they so wish.

A user will NOT be emailed if any of the following are true
- they have already indicated their preference for transferral
- they have received an email on the matter within the current week
- they have already received 3 emails on the matter

The email itself will have two links, one to accept and the other to deny the transferral
of the personal series between accounts. If a user accepts the transferral, a "My Videos"
tool placement will be added to their Vula workspace, and ACLs will be updated (old user
removed, new user added) to reflect the change.
Opting not to transfer the series leaves the current state of affairs as is, and simply
stops further communication on series transferral.